--- chatbot-python-project-data-codes/train_chatbot.py
@@ -1,17 +1,0 @@
-# imports
-import nltk
-from nltk.stem import WordNetLemmatizer
-lemmatizer = WordNetLemmatizer()
-import json
-import pickle
-import numpy as np
-from keras.models import Sequential
-from keras.layers import Dense, Activation, Dropout
-from tensorflow.keras.optimizers import SGD
-import random
-words=[]
-classes = []
-documents = []
-ignore_words = ['?', '!']
-data_file = open('intents.json').read()
-intents = json.loads(data_file)
